metal: int4 P10b — in-register T_scale cast for affine quant kernels
Repays the deferred decision recorded at
INT4_PARITY_PROBES.md:259-273 ("Decision: in-register cast")
that P1-P6 silently violated. Every sampled mlx-community 4bit
checkpoint ships F16 scales/biases; pre-P10b
`set_target_dtype(BF16)` truncated those F16 values to BF16 at
load on the bf16 stack, losing 3 mantissa bits per scale
(10→7). After P10b the kernels read scales/biases through a
dedicated `T_scale` device pointer and cast to `T_act` in
registers — preserving the on-disk precision for the qmv hot
path (which already accumulates in float).

Kernel template split (`shaders/quantized_*.metal`):
- `affine_qmv` / `qmv_fast` / `qmv_quad`,
  `affine_qmm_t` / `qmm_t_splitk` / `qmm_n`,
  `affine_qvm` / `qvm_split_k`,
  `affine_dequantize_b4` / `affine_embed_b4`
  go from `<T,…>` to `<T_act, T_scale,…>`.
- Symbol naming: `affine_<op>_<T_act>_s_<T_scale>_gs_…`.
- Coverage: `half×half` + `bfloat×half` only; the regression
  `bfloat×bfloat` instantiations are removed.

Host dispatcher (`quantized.rs`):
- Add `pub enum ScaleDtype { F16 }` (single variant in P10b;
  forward-compat for P11 mixed-quant).
- Plumb `scale_dtype` through `MetalAffineDequantize::execute`,
  `MetalAffineEmbed::execute`, `MetalAffineQmv::execute`,
  `MetalAffineQmmT::execute`, `MetalAffineQmmN::execute`,
  `MetalAffineQvm::execute`, and their `*_kernel_name` /
  `*_kernel_static_name` helpers.

Lowering (`ferrite-forward/src/interpreter/metal/lowering.rs`):
- Add `scale_dtype_for<W>()` next to `dequant_dtype_for<W>()`,
  returning `ScaleDtype::F16` for every metal model today.
- Thread it through the `AffineQmm` (matvec + qmm_t Standard +
  SplitK arms) and `AffineEmbed` lowering arms.

Loader exclusion (`ferrite-cuda-core/src/weights.rs`,
`ferrite-kernels/src/layers.rs`):
- New `GpuWeights::take_keep_dtype` bypasses `maybe_cast_cpu`
  so on-disk bytes go to the device verbatim.
- `AffineQuantLinear::load` / `AffineQuantEmbedding::load` use
  it for `*.scales` / `*.biases`. Quant scales/biases stay F16
  on the device (also saving ~5MB per Llama-1B vs the cast
  scratch).

CPU reference (`cpu_reference.rs`) is now parametrized on
`<TAct, TScale>` so the bf16-act entry points consume F16
scales (matching production); mirrors the kernel's in-register
T_scale → T_act cast for bit-exact CPU/GPU comparison.

Tests:
- All 38 affine-quant kernel parity tests still pass with the
  new F16-scale × bf16-act inputs (`cpu_reference_test`,
  `quantized_qmv_test`, `quantized_qmm_test`,
  `quantized_qmm_n_test`, `quantized_qvm_test`,
  `quantized_dequantize_test`, `quantized_affine_embed_test`,
  `quantized_splitk_reduce_test`).
- `test_metal_correctness_llama_3_2_1b_mlx_4bit` still passes
  the strict 5-token prefix bar. Prompt 4's first MLX
  divergence moves from ≥90 chars to char 170; prompts 1/2/5
  still match the full 64 tokens. Prompt 7 (the bf16-vs-fp16
  activation drift case, not the regression site) is
  unchanged. Tightening prefix_token_match toward 64 still
  gates on the activation-dtype alignment (P10c+).
- `RUST_LOG=ferrite_cuda_core::weights=debug` shows no more
  `Cast weight: F16 → BF16` lines on `*.scales` / `*.biases`
  or on `embed_tokens.{scales,biases}`. The only remaining
  F16→BF16 casts are the 16 × 2 RMSNorm gains + final norm
  (33 × 2048 elements) — out of P10b's tight scope, deferred
  to a follow-up.

Scoping note: this is the "tight scope" path from the P10b
handoff — only the four `quantized_*.metal` shaders + their
host plumbing. RMSNorm + dense-Linear remain on the loader's
cast path.

+++ vllm-rs/crates/ferrite-metal-kernels/tests/cpu_reference_test.rs
@@ -38,8 +38,9 @@ fn dequant_b4_bf16_scale_bias_apply_per_group() {
     // group1 scale=0.5, bias=-1 → nibbles (1,2,3,4) → -0.5, 0, 0.5, 1
     // Packed bytes pack (lo, hi) → byte = hi<<4 | lo.
     let packed: Vec<u8> = vec![0x43, 0x65, 0x21, 0x43];
-    let scales = vec![half::bf16::from_f32(2.0), half::bf16::from_f32(0.5)];
-    let biases = vec![half::bf16::from_f32(1.0), half::bf16::from_f32(-1.0)];
+    // P10b: bf16-act path takes F16 scales/biases (`T_scale = half`).
+    let scales = vec![half::f16::from_f32(2.0), half::f16::from_f32(0.5)];
+    let biases = vec![half::f16::from_f32(1.0), half::f16::from_f32(-1.0)];
     let mut out = vec![half::bf16::ZERO; 8];
     affine_dequantize_b4_bf16(&packed, &scales, &biases, &mut out, 4);
     let want = [7.0, 9.0, 11.0, 13.0, -0.5, 0.0, 0.5, 1.0];
@@ -61,8 +62,9 @@ fn qmm_t_b4_bf16_all_ones_matches_row_sum() {
     let k = 8;
     let gs = 8;
     let packed = vec![0x11_u8; n * k / 2];
-    let scales = vec![half::bf16::from_f32(1.0); n * k / gs];
-    let biases = vec![half::bf16::ZERO; n * k / gs];
+    // P10b: bf16-act path takes F16 scales/biases.
+    let scales = vec![half::f16::from_f32(1.0); n * k / gs];
+    let biases = vec![half::f16::ZERO; n * k / gs];
     let x: Vec<half::bf16> = (0..(m * k))
         .map(|i| half::bf16::from_f32(i as f32))
         .collect();
@@ -87,8 +89,9 @@ fn qmm_n_b4_bf16_all_ones_matches_row_sum() {
     let k = 8;
     let gs = 4; // qmm_n requires N % gs == 0
     let packed = vec![0x11_u8; k * n / 2];
-    let scales = vec![half::bf16::from_f32(1.0); k * n / gs];
-    let biases = vec![half::bf16::ZERO; k * n / gs];
+    // P10b: bf16-act path takes F16 scales/biases.
+    let scales = vec![half::f16::from_f32(1.0); k * n / gs];
+    let biases = vec![half::f16::ZERO; k * n / gs];
     let x: Vec<half::bf16> = (0..(m * k))
         .map(|i| half::bf16::from_f32(i as f32))
         .collect();
@@ -133,11 +136,12 @@ fn qvm_alias_matches_qmm_n() {
     let k = 8;
     let gs = 4;
     let packed: Vec<u8> = (0..(k * n / 2) as u8).collect();
-    let scales: Vec<half::bf16> = (0..(k * n / gs))
-        .map(|i| half::bf16::from_f32(0.01 + i as f32 * 0.001))
+    // P10b: bf16-act path takes F16 scales/biases.
+    let scales: Vec<half::f16> = (0..(k * n / gs))
+        .map(|i| half::f16::from_f32(0.01 + i as f32 * 0.001))
         .collect();
-    let biases: Vec<half::bf16> = (0..(k * n / gs))
-        .map(|i| half::bf16::from_f32(-0.05 + i as f32 * 0.002))
+    let biases: Vec<half::f16> = (0..(k * n / gs))
+        .map(|i| half::f16::from_f32(-0.05 + i as f32 * 0.002))
         .collect();
     let x: Vec<half::bf16> = (0..(m * k))
         .map(|i| half::bf16::from_f32((i as f32) * 0.1))
